Fix and improve prefiltering
Moreover, the style of applyBlocks in prefiltering deviated from Duncan's original definition. When I started to implement prefiltering, I realized that this was not necessary and I started to make the style more consistent; but then in trying to prove why the new definition was correct, I realized that it was, in fact, _not_ correct in the presence of rollbacks. This is now fixed.